+++ OneBlink.SDK/models/FormSubmissionEventConfigurationMapping.cs
@@ -20,7 +20,7 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public Guid formElementId
+        public Guid? formElementId
         {
             get; set;
         }